Fixed pngs not supporting transparency for the dealers logo

+++ application/modules/admin/models/Dealer.php
@@ -132,21 +132,26 @@ class Admin_Model_Dealer extends My_Model_Abstract
     /**
      * Gets the path (relative to the public path) of the dealer report logo
      *
+     * @param bool $recreate Whether or not the file should be recreated if it exists
+     *
      * @return bool|string
      */
-    public function getDealerLogoImageFile ()
+    public function getDealerLogoImageFile ($recreate = false)
     {
         $publicFilePath = false;
         if ($this->dealerLogoImageId > 0)
         {
             $publicFilePath = '/downloads/dealer-' . $this->id . '-ReportLogo.png';
             $filePath       = PUBLIC_PATH . $publicFilePath;
-            if (file_exists($filePath))
+            if (!file_exists($filePath) || $recreate)
             {
-                @unlink($filePath);
+                if (file_exists($filePath))
+                {
+                    @unlink($filePath);
+                }
+                $image = Admin_Model_Mapper_Image::getInstance()->find($this->dealerLogoImageId);
+                file_put_contents($filePath, base64_decode($image->image));
             }
-            $image = Admin_Model_Mapper_Image::getInstance()->find($this->dealerLogoImageId);
-            file_put_contents($filePath, base64_decode($image->image));
         }
 
         return $publicFilePath;
